refactor(core): extract offscreen capture pipeline into capture.ts
