+++ packages/core/src/app/lib/export-pptx.ts
@@ -1,27 +1,11 @@
-import { createElement } from 'react';
-import { createRoot, type Root } from 'react-dom/client';
-import { designToCssVars } from './design';
+import { capturePagesAsPng } from './capture';
 import { downloadBlob } from './download';
-import { SlidePageProvider } from './page-context';
-import { isFrameAnimationSettled, waitForDataWaitfor, waitForFonts } from './print-ready';
+import { FORMAT_PRESETS } from './formats';
 import type { SlideModule } from './sdk';
 
-const SLIDE_W = 1920;
-const SLIDE_H = 1080;
 // 16:9 widescreen in English Metric Units (914400 EMU per inch → 13.333in × 7.5in).
 const EMU_W = 12192000;
 const EMU_H = 6858000;
-const CAPTURE_PIXEL_RATIO = 2;
-
-const ANIMATION_TIMEOUT_MS = 15_000;
-const POLL_INTERVAL_MS = 100;
-
-const CAPTURE_CLASS = 'os-pptx-capture';
-const CAPTURE_STYLE_ID = 'os-pptx-capture-style';
-// Properties intro animations drive from a hidden start state to a visible end
-// state. We read them back once settled and pin them inline so the capture clone
-// can't re-run the keyframes from their invisible 0% frame (see freezeForCapture).
-const FROZEN_PROPS = ['opacity', 'transform', 'filter', 'clip-path'] as const;
 
 export type PptxExportProgress = {
   phase: 'processing' | 'generating' | 'done';
@@ -43,91 +27,15 @@ export async function exportSlideAsImagePptx(
   const total = pages.length;
   onProgress?.({ phase: 'processing', current: 0, total, percent: 0 });
 
-  const container = document.createElement('div');
-  container.className = CAPTURE_CLASS;
-  container.setAttribute('aria-hidden', 'true');
-  Object.assign(container.style, {
-    position: 'fixed',
-    left: '-99999px',
-    top: '0',
-    pointerEvents: 'none',
-  });
-  document.body.appendChild(container);
-
-  // html-to-image clones each frame and copies its computed style — including the
-  // intro animation — into the clone, which then re-runs the keyframes from their
-  // hidden 0% frame in the rasterised SVG. Fast-forward every animation to its end
-  // frame in the live DOM (a large negative delay lands past a 1ms duration, so
-  // even pseudo-elements paint their final state on the first frame).
-  const captureStyle = document.createElement('style');
-  captureStyle.id = CAPTURE_STYLE_ID;
-  captureStyle.textContent = `.${CAPTURE_CLASS} *, .${CAPTURE_CLASS} *::before, .${CAPTURE_CLASS} *::after {
-    animation-delay: -1s !important;
-    animation-duration: 1ms !important;
-    animation-iteration-count: 1 !important;
-    animation-fill-mode: forwards !important;
-    transition: none !important;
-  }`;
-  document.head.appendChild(captureStyle);
-
-  const designVars = slide.design ? designToCssVars(slide.design) : null;
-
-  const reactRoots: Root[] = [];
-  const frames: HTMLElement[] = [];
-  for (let i = 0; i < pages.length; i++) {
-    const Page = pages[i];
-    if (!Page) continue;
-    const host = document.createElement('div');
-    host.setAttribute('data-osd-canvas', '');
-    host.style.width = `${SLIDE_W}px`;
-    host.style.height = `${SLIDE_H}px`;
-    host.style.overflow = 'hidden';
-    host.style.background = '#fff';
-    if (designVars) {
-      for (const [k, v] of Object.entries(designVars)) host.style.setProperty(k, v);
-    }
-    container.appendChild(host);
-    frames.push(host);
-    const r = createRoot(host);
-    r.render(
-      createElement(SlidePageProvider, { index: i, total: pages.length }, createElement(Page)),
-    );
-    reactRoots.push(r);
-  }
-  // Yield once so React commits all pages and intro animations actually start.
-  await nextPaint();
-
   try {
-    await waitForFonts();
-
-    const deadline = performance.now() + ANIMATION_TIMEOUT_MS;
-    while (performance.now() < deadline) {
-      const settled = frames.every((frame) => isFrameAnimationSettled(frame));
-      if (settled) break;
-      await sleep(POLL_INTERVAL_MS);
-    }
-    await waitForDataWaitfor(container);
-
-    const { toBlob } = await import('html-to-image');
-    const images: Uint8Array[] = [];
-    for (let i = 0; i < frames.length; i++) {
-      freezeForCapture(frames[i]);
-      const blob = await toBlob(frames[i], {
-        width: SLIDE_W,
-        height: SLIDE_H,
-        pixelRatio: CAPTURE_PIXEL_RATIO,
-        backgroundColor: '#ffffff',
-        cacheBust: true,
-      });
-      if (!blob) throw new Error(`failed to capture page ${i + 1}`);
-      images.push(new Uint8Array(await blob.arrayBuffer()));
+    const images = await capturePagesAsPng(slide, FORMAT_PRESETS.slide, (captured) => {
       onProgress?.({
         phase: 'processing',
-        current: i + 1,
+        current: captured,
         total,
-        percent: Math.min(95, ((i + 1) / total) * 95),
+        percent: Math.min(95, (captured / total) * 95),
       });
-    }
+    });
 
     onProgress?.({ phase: 'generating', current: total, total, percent: 98 });
     const pptx = await buildImagePptx(images);
@@ -139,23 +47,6 @@ export async function exportSlideAsImagePptx(
     );
   } finally {
     onProgress?.({ phase: 'done', current: total, total, percent: 100 });
-    for (const r of reactRoots) r.unmount();
-    container.remove();
-    captureStyle.remove();
-  }
-}
-
-// Pin each element's settled visual state inline and remove its animation so the
-// clone html-to-image rasterises renders the final frame instead of replaying the
-// (initially invisible) keyframes. Pseudo-elements are handled by CAPTURE_STYLE_ID.
-function freezeForCapture(root: HTMLElement): void {
-  for (const el of root.querySelectorAll<HTMLElement>('*')) {
-    const cs = getComputedStyle(el);
-    for (const prop of FROZEN_PROPS) {
-      el.style.setProperty(prop, cs.getPropertyValue(prop), 'important');
-    }
-    el.style.setProperty('animation', 'none', 'important');
-    el.style.setProperty('transition', 'none', 'important');
   }
 }
 
@@ -253,21 +144,4 @@ function slideRelsXml(idx: number): string {
 
 function themeXml(): string {
   return `${XML_DECL}<a:theme xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" name="Office Theme"><a:themeElements><a:clrScheme name="Office"><a:dk1><a:sysClr val="windowText" lastClr="000000"/></a:dk1><a:lt1><a:sysClr val="window" lastClr="FFFFFF"/></a:lt1><a:dk2><a:srgbClr val="44546A"/></a:dk2><a:lt2><a:srgbClr val="E7E6E6"/></a:lt2><a:accent1><a:srgbClr val="4472C4"/></a:accent1><a:accent2><a:srgbClr val="ED7D31"/></a:accent2><a:accent3><a:srgbClr val="A5A5A5"/></a:accent3><a:accent4><a:srgbClr val="FFC000"/></a:accent4><a:accent5><a:srgbClr val="5B9BD5"/></a:accent5><a:accent6><a:srgbClr val="70AD47"/></a:accent6><a:hlink><a:srgbClr val="0563C1"/></a:hlink><a:folHlink><a:srgbClr val="954F72"/></a:folHlink></a:clrScheme><a:fontScheme name="Office"><a:majorFont><a:latin typeface="Calibri Light"/><a:ea typeface=""/><a:cs typeface=""/></a:majorFont><a:minorFont><a:latin typeface="Calibri"/><a:ea typeface=""/><a:cs typeface=""/></a:minorFont></a:fontScheme><a:fmtScheme name="Office"><a:fillStyleLst><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:gradFill rotWithShape="1"><a:gsLst><a:gs pos="0"><a:schemeClr val="phClr"><a:lumMod val="110000"/><a:satMod val="105000"/><a:tint val="67000"/></a:schemeClr></a:gs><a:gs pos="50000"><a:schemeClr val="phClr"><a:lumMod val="105000"/><a:satMod val="103000"/><a:tint val="73000"/></a:schemeClr></a:gs><a:gs pos="100000"><a:schemeClr val="phClr"><a:lumMod val="105000"/><a:satMod val="109000"/><a:tint val="81000"/></a:schemeClr></a:gs></a:gsLst><a:lin ang="5400000" scaled="0"/></a:gradFill><a:gradFill rotWithShape="1"><a:gsLst><a:gs pos="0"><a:schemeClr val="phClr"><a:satMod val="103000"/><a:lumMod val="102000"/><a:tint val="94000"/></a:schemeClr></a:gs><a:gs pos="50000"><a:schemeClr val="phClr"><a:satMod val="110000"/><a:lumMod val="100000"/><a:shade val="100000"/></a:schemeClr></a:gs><a:gs pos="100000"><a:schemeClr val="phClr"><a:lumMod val="99000"/><a:satMod val="120000"/><a:shade val="78000"/></a:schemeClr></a:gs></a:gsLst><a:lin ang="5400000" scaled="0"/></a:gradFill></a:fillStyleLst><a:lnStyleLst><a:ln w="6350" cap="flat" cmpd="sng" algn="ctr"><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:prstDash val="solid"/><a:miter lim="800000"/></a:ln><a:ln w="12700" cap="flat" cmpd="sng" algn="ctr"><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:prstDash val="solid"/><a:miter lim="800000"/></a:ln><a:ln w="19050" cap="flat" cmpd="sng" algn="ctr"><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:prstDash val="solid"/><a:miter lim="800000"/></a:ln></a:lnStyleLst><a:effectStyleLst><a:effectStyle><a:effectLst/></a:effectStyle><a:effectStyle><a:effectLst/></a:effectStyle><a:effectStyle><a:effectLst/></a:effectStyle></a:effectStyleLst><a:bgFillStyleLst><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:solidFill><a:schemeClr val="phClr"><a:tint val="95000"/><a:satMod val="170000"/></a:schemeClr></a:solidFill><a:gradFill rotWithShape="1"><a:gsLst><a:gs pos="0"><a:schemeClr val="phClr"><a:tint val="93000"/><a:satMod val="150000"/><a:shade val="98000"/><a:lumMod val="102000"/></a:schemeClr></a:gs><a:gs pos="50000"><a:schemeClr val="phClr"><a:tint val="98000"/><a:satMod val="130000"/><a:shade val="90000"/><a:lumMod val="103000"/></a:schemeClr></a:gs><a:gs pos="100000"><a:schemeClr val="phClr"><a:shade val="63000"/><a:satMod val="120000"/></a:schemeClr></a:gs></a:gsLst><a:lin ang="5400000" scaled="0"/></a:gradFill></a:bgFillStyleLst></a:fmtScheme></a:themeElements></a:theme>`;
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function nextPaint(): Promise<void> {
-  return new Promise((resolve) => {
-    let settled = false;
-    const settle = () => {
-      if (settled) return;
-      settled = true;
-      resolve();
-    };
-    requestAnimationFrame(settle);
-    setTimeout(settle, 50);
-  });
 }
